+++ apps/server/src/api/products/dataexchange/dataexchange.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common"
+import { DataexchangeService } from "./dataexchange.service"
+import { DataexchangeController } from "./dataexchange.controller"
+
+@Module({
+  controllers: [DataexchangeController],
+  providers: [DataexchangeService],
+})
+export class DataexchangeModule { }