+++ apps/server/src/api/products/dataexchange/models/dataset.model.ts
@@ -1,0 +1,4 @@
+import { dataexchangeMongoDbConn } from "src/lib/db-connect"
+import { DatasetSchema } from "../schemas/dataset.schema"
+
+export const DatasetModel = dataexchangeMongoDbConn.model("dataset", DatasetSchema)