+++ apps/server/src/api/products/dataexchange/models/metadata.model.ts
@@ -1,0 +1,4 @@
+import { dataexchangeMongoDbConn } from "src/lib/db-connect"
+import { MetadataSchema } from "../schemas/metadata.schema"
+
+export const MetaDataModel = dataexchangeMongoDbConn.model("metadata", MetadataSchema)