+++ apps/server/src/api/products/dataexchange/queries/find-categories.query.ts
@@ -1,0 +1,8 @@
+import { MetaDataModel } from "../models/metadata.model"
+
+export async function findDistinctCategories() {
+  const filterCategories = await MetaDataModel.find().distinct("category")
+  filterCategories.push("All")
+  filterCategories.sort()
+  return filterCategories
+}