+++ apps/server/src/api/products/dataexchange/queries/find-data.query.ts
@@ -1,0 +1,6 @@
+import { DatasetModel } from "../models/dataset.model"
+
+export async function findDatasetDataById(datasetId: string) {
+  const dataset = await DatasetModel.findOne({ datasetRelationId: datasetId })
+  return dataset
+}