+++ apps/server/src/api/products/dataexchange/queries/find-datasets.query.ts
@@ -1,0 +1,15 @@
+import { MetaDataModel } from "../models/metadata.model"
+
+export async function findDatasets(searchQuery: string, selectedFilterCategory: string, selectedSortOption: string, offset: number, limit: number) {
+  let datasets = await MetaDataModel.find({
+    $or: [
+      { name: { $regex: searchQuery, $options: "i" } },
+      { description: { $regex: searchQuery, $options: "i" } }
+    ],
+    category: { $regex: selectedFilterCategory }
+  }).sort(selectedSortOption)
+    .skip(offset)
+    .limit(limit)
+
+  return datasets
+}