+++ apps/server/src/api/products/products.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common"
-import { DatalakeModule } from "./datalake/datalake.module"
+import { DataexchangeModule } from "./dataexchange/dataexchange.module"
 import { SwapModule } from "./swap/swap.module"
 import { WalletModule } from "./wallet/wallet.module"
 import { NftstudioModule } from "./nftstudio/nftstudio.module"
@@ -9,7 +9,7 @@ import { CopilotModule } from "./copilot/copilot.module"
 import { AnalyticsModule } from './analytics/analytics.module'
 
 @Module({
-  imports: [DatalakeModule, SwapModule, WalletModule, NftstudioModule, FabricModule, LedgerscanModule, CopilotModule, AnalyticsModule],
+  imports: [DataexchangeModule, SwapModule, WalletModule, NftstudioModule, FabricModule, LedgerscanModule, CopilotModule, AnalyticsModule],
 })
 
 export class ProductsModule { }