+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Added type safety for dispatcher in context
 * Deprecation of Insights 
 * Introduction of Analytics with new CQRS architecture
+* Deprecation of Datalake 
+* Introduction of Data Exchange with new CQRS architecture
 
 ### v9.3.1
 * Minor bug fixes & performance improvements